app/requirements_web.txt
uvicorn==0.30.1
fastapi==0.115.8
SQLAlchemy==2.0.31
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.1.10
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
sqlmodel
starlette>=0.45.0,<0.46.0
sqlalchemy_utils
python-jose==3.3.0
pika==1.3.2
opyoid==3.0.1
pytest==7.4.3
pytest-mock==3.14.0
httpx==0.25.1
python-multipart==0.0.15
urllib3<2.0
python-dotenv==1.0.0
streamlit==1.29.0